refactor 取得したドメインに合わせてパッケージ変更
