+++ src/main/java/click/divichart/configuration/WebSecurityConfig.java
@@ -1,4 +1,4 @@
-package jp.ne.divichart.configuration;
+package click.divichart.configuration;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
