+++ src/main/java/click/divichart/controller/BarChartController.java
@@ -1,6 +1,6 @@
-package jp.ne.divichart.controller;
+package click.divichart.controller;
 
-import jp.ne.divichart.service.PieChartService;
+import click.divichart.service.BarChartService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,16 +11,16 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping({"/", "/pieChart"})
-public class PieChartController {
+@RequestMapping("/barChart")
+public class BarChartController {
 
-    private static final Logger log = LoggerFactory.getLogger(PieChartController.class);
+    private static final Logger log = LoggerFactory.getLogger(BarChartController.class);
     @Autowired
-    PieChartService service;
+    BarChartService service;
 
     @GetMapping
     public String index(Model model, @ModelAttribute("targetYear") String targetYear) {
-        log.debug("配当割合グラフ表示");
+        log.debug("月別配当グラフ表示");
 
         String[] recentYears = service.getRecentYears();
         model.addAttribute("recentYears", recentYears);
@@ -28,10 +28,10 @@ public class PieChartController {
         if (targetYear.isEmpty() || service.isNotYear(targetYear)) targetYear = recentYears[0];
         model.addAttribute("targetYear", targetYear);
 
-        String[] chartData = service.getChartData(targetYear);
+        String chartData = service.getChartData(targetYear);
         model.addAttribute("chartData", chartData);
 
-        return "pieChart";
+        return "barChart";
     }
 
 }
