+++ src/main/java/click/divichart/controller/LineChartController.java
@@ -1,6 +1,6 @@
-package jp.ne.divichart.controller;
+package click.divichart.controller;
 
-import jp.ne.divichart.service.BarChartService;
+import click.divichart.service.LineChartService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,28 +10,29 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-@Controller
-@RequestMapping("/barChart")
-public class BarChartController {
 
-    private static final Logger log = LoggerFactory.getLogger(BarChartController.class);
+@Controller
+@RequestMapping("/lineChart")
+public class LineChartController {
+
+    private static final Logger log = LoggerFactory.getLogger(LineChartController.class);
+
     @Autowired
-    BarChartService service;
+    LineChartService service;
 
     @GetMapping
     public String index(Model model, @ModelAttribute("targetYear") String targetYear) {
-        log.debug("月別配当グラフ表示");
+        log.debug("累計配当グラフ表示");
 
         String[] recentYears = service.getRecentYears();
-        model.addAttribute("recentYears", recentYears);
-
         if (targetYear.isEmpty() || service.isNotYear(targetYear)) targetYear = recentYears[0];
-        model.addAttribute("targetYear", targetYear);
 
         String chartData = service.getChartData(targetYear);
         model.addAttribute("chartData", chartData);
 
-        return "barChart";
+        model.addAttribute("targetYear", targetYear);
+        model.addAttribute("recentYears", recentYears);
+        return "lineChart";
     }
 
 }
