+++ src/main/java/click/divichart/controller/PieChartController.java
@@ -1,6 +1,6 @@
-package jp.ne.divichart.controller;
+package click.divichart.controller;
 
-import jp.ne.divichart.service.LineChartService;
+import click.divichart.service.PieChartService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,29 +10,28 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-
 @Controller
-@RequestMapping("/lineChart")
-public class LineChartController {
+@RequestMapping({"/", "/pieChart"})
+public class PieChartController {
 
-    private static final Logger log = LoggerFactory.getLogger(LineChartController.class);
-
+    private static final Logger log = LoggerFactory.getLogger(PieChartController.class);
     @Autowired
-    LineChartService service;
+    PieChartService service;
 
     @GetMapping
     public String index(Model model, @ModelAttribute("targetYear") String targetYear) {
-        log.debug("累計配当グラフ表示");
+        log.debug("配当割合グラフ表示");
 
         String[] recentYears = service.getRecentYears();
-        if (targetYear.isEmpty() || service.isNotYear(targetYear)) targetYear = recentYears[0];
+        model.addAttribute("recentYears", recentYears);
 
-        String chartData = service.getChartData(targetYear);
+        if (targetYear.isEmpty() || service.isNotYear(targetYear)) targetYear = recentYears[0];
+        model.addAttribute("targetYear", targetYear);
+
+        String[] chartData = service.getChartData(targetYear);
         model.addAttribute("chartData", chartData);
 
-        model.addAttribute("targetYear", targetYear);
-        model.addAttribute("recentYears", recentYears);
-        return "lineChart";
+        return "pieChart";
     }
 
 }
