+++ src/test/java/click/divichart/service/LineChartServiceTest.java
@@ -1,4 +1,4 @@
-package jp.ne.divichart.service;
+package click.divichart.service;
